+++ packages/capture/src/index.ts
@@ -11,6 +11,7 @@
  */
 export * from "./process";
 export * from "./browser";
+export * from "./determinism";
 export * from "./settle";
 export * from "./cursor";
 export * from "./native";